@@ -12,25 +12,25 @@
     "build:watch": "ng build demo --watch --configuration development"
   },
   "dependencies": {
-    "@angular/animations": "^16.1.3",
-    "@angular/common": "^16.1.3",
-    "@angular/compiler": "^16.1.3",
-    "@angular/core": "^16.1.3",
-    "@angular/forms": "^16.1.3",
-    "@angular/platform-browser": "^16.1.3",
-    "@angular/platform-browser-dynamic": "^16.1.3",
-    "@angular/router": "^16.1.3",
+    "@angular/animations": "^19.2.5",
+    "@angular/common": "^19.2.5",
+    "@angular/compiler": "^19.2.5",
+    "@angular/core": "^19.2.5",
+    "@angular/forms": "^19.2.5",
+    "@angular/platform-browser": "^19.2.5",
+    "@angular/platform-browser-dynamic": "^19.2.5",
+    "@angular/router": "^19.2.5",
     "ngx-bootstrap": "^10.2.0",
     "rxjs": "^7.8.1",
     "tslib": "^2.6.0",
-    "zone.js": "~0.13.1"
+    "zone.js": "0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.3",
-    "@angular/cli": "^16.1.3",
-    "@angular/compiler-cli": "^16.1.3",
+    "@angular-devkit/build-angular": "^19.2.5",
+    "@angular/cli": "^19.2.5",
+    "@angular/compiler-cli": "^19.2.5",
     "@types/estree": "^1.0.1",
-    "@types/node": "^20.3.3",
-    "typescript": "~5.1.3"
+    "@types/node": "^22.15.2",
+    "typescript": "5.8.3"
   }
 }
