@@ -1,7 +1,7 @@
 {
   "name": "ngx-ui-scroll",
   "description": "Virtual scrolling library for Angular applications",
-  "version": "3.2.2",
+  "version": "4.0.0",
   "license": "MIT",
   "author": "Denis Hilt <dhilt.public@gmail.com>",
   "homepage": "https://github.com/dhilt/ngx-ui-scroll",
@@ -16,15 +16,15 @@
     "build:watch": "ng build scroller --watch"
   },
   "peerDependencies": {
-    "@angular/common": ">=12.0.0",
-    "@angular/core": ">=12.0.0",
+    "@angular/common": ">=17.0.0",
+    "@angular/core": ">=17.0.0",
     "vscroll": "1.6.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.0.0",
-    "@angular/cli": "^12.0.0",
-    "@angular/compiler": "^12.0.0",
-    "ng-packagr": "^12.0.0",
-    "typescript": "~4.3.0"
+    "@angular-devkit/build-angular": "^19.0.0",
+    "@angular/cli": "^19.0.0",
+    "@angular/compiler": "^19.0.0",
+    "ng-packagr": "^19.0.0",
+    "typescript": "5.8.2"
   }
 }
