@@ -7,21 +7,23 @@
     "test": "ng test scroller"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.3",
-    "@angular/cli": "^16.1.3",
-    "@angular/compiler": "^16.1.3",
-    "@angular/core": "^16.1.3",
-    "@angular/platform-browser": "^16.1.3",
-    "@angular/platform-browser-dynamic": "^16.1.3",
-    "@types/jasmine": "^4.3.5",
-    "@types/node": "^20.3.3",
-    "jasmine-core": "^5.0.1",
-    "karma": "^6.4.2",
+    "@angular-devkit/build-angular": "^19.2.5",
+    "@angular/cli": "^19.2.5",
+    "@angular/compiler": "^19.2.5",
+    "@angular/compiler-cli": "^19.2.5",
+    "@angular/core": "^19.2.5",
+    "@angular/platform-browser": "^19.2.5",
+    "@angular/platform-browser-dynamic": "^19.2.5",
+    "@types/jasmine": "^5.1.7",
+    "@types/node": "^22.15.2",
+    "jasmine-core": "^5.6.0",
+    "karma": "^6.4.4",
     "karma-chrome-launcher": "^3.2.0",
-    "karma-firefox-launcher": "^2.1.2",
+    "karma-firefox-launcher": "^2.1.3",
     "karma-jasmine": "^5.1.0",
-    "puppeteer": "^20.7.4",
-    "typescript": "~5.1.3",
-    "zone.js": "~0.13.1"
+    "karma-spec-reporter": "^0.0.36",
+    "puppeteer": "^24.7.2",
+    "typescript": "5.8.3",
+    "zone.js": "0.15.0"
   }
 }
