@@ -10,7 +10,8 @@ import { filter } from 'rxjs/operators';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html'
+  templateUrl: './app.component.html',
+  standalone: false
 })
 export class AppComponent implements AfterViewInit, OnDestroy {
   hasLayout = true;
