@@ -4,7 +4,8 @@ import { DemoSources } from '../interfaces';
 
 @Component({
   selector: 'app-demo-sources',
-  templateUrl: './demo-sources.component.html'
+  templateUrl: './demo-sources.component.html',
+  standalone: false
 })
 export class DemoSourcesComponent {
   @Input() sources!: DemoSources;
