@@ -4,7 +4,8 @@ import { IDemo } from '../../routes';
 
 @Component({
   selector: 'app-demo-title',
-  templateUrl: './demo-title.component.html'
+  templateUrl: './demo-title.component.html',
+  standalone: false
 })
 export class DemoTitleComponent {
   @Input() config!: IDemo;
