@@ -4,7 +4,8 @@ import { demos } from '../routes';
 
 @Component({
   selector: 'app-nav',
-  templateUrl: './nav.component.html'
+  templateUrl: './nav.component.html',
+  standalone: false
 })
 export class NavComponent {
   demos = demos;
