@@ -1,12 +1,7 @@
-import {
-  makeTest,
-  ItFunc,
-  TestBedConfig,
-  ItFuncConfig
-} from '../scaffolding/runner';
-import { INVALID_DATASOURCE_PREFIX, Settings } from '../miscellaneous/vscroll';
+import { makeTest, ItFunc, TestBedConfig } from '../scaffolding/runner';
+import { Settings } from '../miscellaneous/vscroll';
 
-const configClassPass: TestBedConfig<void, false> = {
+const configClass: TestBedConfig<void, false> = {
   datasourceClass: class {
     settings: Settings;
     constructor() {
@@ -15,80 +10,9 @@ const configClassPass: TestBedConfig<void, false> = {
     get(_offset: number, _count: number) {
       return [];
     }
-    reset() {}
+    reset() { }
   }
 };
-
-interface ICustomClassFail {
-  title: string;
-  error?: string;
-  errorStart?: string;
-}
-
-const configListClassFail: TestBedConfig<ICustomClassFail, false>[] = [
-  {
-    datasourceClass: 'invalid' as unknown as { new (): void },
-    toThrow: true,
-    custom: {
-      title: 'should throw exception (datasource is not a constructor)',
-      error: 'datasource is not a constructor'
-    }
-  },
-  {
-    datasourceClass: class {
-      settings: Settings;
-      constructor() {
-        this.settings = {};
-      }
-    },
-    toThrow: true,
-    custom: {
-      title: 'should throw exception (no get)',
-      errorStart: INVALID_DATASOURCE_PREFIX
-    }
-  },
-  {
-    datasourceClass: class {
-      settings: Settings;
-      get: boolean;
-      constructor() {
-        this.settings = {};
-        this.get = true;
-      }
-    },
-    toThrow: true,
-    custom: {
-      title: 'should throw exception (get is not a function)',
-      errorStart: INVALID_DATASOURCE_PREFIX
-    }
-  },
-  {
-    datasourceClass: class {
-      settings: Settings;
-      constructor() {
-        this.settings = {};
-      }
-      get(offset: number) {
-        return ++offset;
-      }
-    },
-    toThrow: true,
-    custom: {
-      title: 'should throw exception (get has less than 2 arguments)',
-      errorStart: INVALID_DATASOURCE_PREFIX
-    }
-  }
-];
-
-const shouldFail: ItFuncConfig<ICustomClassFail, false> =
-  config => error => done => {
-    if (config.custom.error) {
-      expect(error as unknown).toBe(config.custom.error);
-    } else if (config.custom.errorStart) {
-      expect(error.toString().startsWith(config.custom.errorStart)).toBeTrue();
-    }
-    done();
-  };
 
 const shouldWork: ItFunc = misc => async done => {
   await misc.relaxNext();
@@ -107,21 +31,11 @@ const shouldNotFetch: ItFunc = misc => async done => {
 };
 
 describe('Datasource Class', () => {
-  describe('Pass', () =>
-    makeTest<void, false>({
-      config: configClassPass,
-      title: 'should pass',
-      it: () => done => done()
-    }));
-
-  describe('Fail', () =>
-    configListClassFail.forEach(config =>
-      makeTest<ICustomClassFail, false>({
-        config: config,
-        title: config.custom.title,
-        it: shouldFail(config)
-      })
-    ));
+  makeTest<void, false>({
+    config: configClass,
+    title: 'should pass',
+    it: () => done => done()
+  });
 });
 
 describe('Datasource Get', () => {
